requirements.txt
python-docx>=0.8.10
python-pptx>=0.6.18
Pillow>=6.0.0
pydantic>=2.9.2
rapidfuzz>=0.10.0
scipy>=1.14.1
numpy>=2.1.3
